renamed coursematrix and refactored into a class
